Fixed test visitor runtime issue.
The extra test changed the meaning of mml::Shape in its compilation
unit. While this was ok when each test was a different program, when all
the tests are linked together it can result in *strange* behaviour at
runtime... The solution was therefore to build a second test binary for
this test.